A lot is different
variable names in DB and program are different <ClassnameId> etc
I have made a CMSContext file that connects to the db
I have updated the web.config to connect to the db
The migrations folder and configurations for the database design are now
in Mappings
The database has its relationships and can cascade on delete
I am working on the items part so dont worry about that not loading.
I have changed the items model and view and added custom scripts folder
that are my scripts seperate from the ones the web app generated. They
are to dynamically load and update the items page.